Change components to useState
The existing code contained a flawless use of the `class` react components, which were prevalent in 2016-2019. Now in 2020 the React ecosystem is shifting (gradually) to always use components that are functions, and use hooks such as `useState` to handle a component's state and `useEffect` to handle transitions previously managed by confusing methods such as `compnentDidMount` and the like. Accordingly, this commit shows what transitioning from old to new looks like.

Main change:
* Update the `<App />` component away from the recently-legacy ES6 class syntax in favor of the new 'hooks' syntax

Other changes:
* Update some variable names to be slightly more expressive and/or conventional (e.g. `handleClose` rather than `close`: the `handle` indicates we're responding to a browser event
* Destructure the props being passed into `<Modal />`
* Generate a `yarn.lock` file as a result of running `yarn` to yarn-install the package dependencies specified in `package.json`